fix screen auth regester
