--- Атестация промежуточная/src/comp/Form/index.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable max-len */
-import React from 'react';
-
-import Connecter from './connected';
-
-const FormBlock = () => {
-  return (
-    <Connecter />
-  );
-};
-
-export default FormBlock;
