--- Атестация промежуточная/src/comp/Form/styled.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const FormRegister = styled.div`
-height: 80vh;
-margin:auto;
-display:flex;
-align-items: center;
-justify-content: center;
-`;
-
-
-export default FormRegister;
