+++ Атестация промежуточная/src/text/index.tsx
@@ -6,18 +6,35 @@ export default {
   'MyProfile': 'Личный профиль',
   'Title': 'Delta World @ 1970-2077',
   'WHITE': 'Cветлая тема',
+  'authButton': 'Войти',
+  'authChange': 'Уже есть аккаунт войти',
+  'authorization': 'Авторизоваться',
   'birthday': 'Дата рождения',
+  'birthdayError': 'birthdayError',
   'createPost': 'Написать пост',
   'email': 'E-mail',
+  'emailError': 'emailError',
   'exit': 'Выход',
+  'female': 'female',
+  'formErrors': 'formErrors',
   'gender': 'Пол',
+  'genderError': 'genderError',
   'header': 'Delta World',
   'id': 'id',
+  'male': 'male',
   'mobile': 'Номер Телефона',
+  'name': 'name',
+  'nameError': 'nameError',
+  'phone': 'phone',
+  'phoneError': 'phoneError',
+  'placeholderBirthday': 'placeholderBirthday',
+  'placeholderEmail': 'placeholderEmail',
   'placeholderId': 'placeholderId',
+  'placeholderName': 'placeholderName',
+  'placeholderPhone': 'placeholderPhone',
   'post': 'Посты',
-  'register': 'Авторизоваться',
-  'registerButton': 'Войти',
+  'register': 'Регистрация',
+  'registerButton': 'Зарегистрировать',
   'registerChange': 'Зарегистрировать аккаунт',
   'signIn': 'Вход',
   'signUn': 'Регистрация',
